requirements.txt
mucsi96.publish-tools == 34.0.0